+++ ergo-lib/src/ast.rs
@@ -1,14 +1,19 @@
 //! AST for ErgoTree
 
+pub(crate) mod apply;
 pub(crate) mod bin_op;
+pub(crate) mod block;
 pub(crate) mod coll_fold;
 pub(crate) mod expr;
 pub(crate) mod extract_reg_as;
+pub(crate) mod func_value;
 pub(crate) mod global_vars;
 pub(crate) mod method_call;
 pub(crate) mod option_get;
 pub(crate) mod predef_func;
 pub(crate) mod property_call;
+pub(crate) mod val_def;
+pub(crate) mod val_use;
 
 pub mod constant;
 pub mod value;